--- test/test_install_file.py
@@ -1,8 +1,0 @@
-import unittest
-
-from test import base
-
-
-class TestInstallFile(base.TestBase, unittest.TestCase):
-
-    pass
